moved API specific parts from SubmodelAPI
