+++ src/IO.Swagger.V1RC03/Controllers/SubmodelAPI.cs
@@ -17,7 +17,6 @@ using System.ComponentModel.DataAnnotations;
 using IO.Swagger.V1RC03.Attributes;
 
 using Microsoft.AspNetCore.Authorization;
-using IO.Swagger.V1RC03.Models;//remove later
 using AasCore.Aas3_0_RC02;
 using IO.Swagger.V1RC03.ApiModel;
